requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.1
wheel==0.44.0